refactor: separate scanner core
